This is probs the end of Quadtree testing
the thing didn't go as well as I hoped it would.
Tbh, maybe the amount of rays needed to test up to the detailed lvl of
image exceeds the basic quad option. (SHAME)

Will keep this one in the back of my mind and maybe copy it back to
other tries as well! Until then cy@!